fix: splitting key functions into new service
